requirements.txt
keras==2.6.0
numpy==1.20.3
pandas==1.3.4
Pillow==9.1.0
scikit_learn==1.0.2
tensorflow==2.6.0
